Refactor: split the base class contract into four focused files
SubjectBaseContract.cs had grown to 953 lines. It held four separable
concerns: which ancestor is a subject, what the generator emits, whether a
base can host a generated subclass, and which members on the chain collide
with the generated half.

Pure move, no behaviour change. The generated output is byte identical
across all 278 emitted subject files.

The shared base chain lookups move to SymbolExtensions, where they sit
beside the other symbol helpers and no longer carry subject specific
parameter names.
